Add source code for the images of last post
The pretty git pictures were generated by a JS package called
gitgraph. It generates SVG elements embeddable in a webpage.
It doesn’t explicitly support saving the output to a file,
but it’s reasonably straightforward to do: just open
the htmls in the browser, enable devtools in Chrome, copy
the svg element’s outerhtml, save, and hand-edit (add the missing
xmlns and xmlns:xlink).