Refactor output directory structure
make build now produces an out/ directory containing three
subdirectories, each corresponding to a domain. All assets
are copied to each subdirectory independently.

This is a bit suboptimal (it’d be best if each domain only
had the assets it needs), but will do for now. The benefits
are that server configuration can be simpler, less manual
config is necessary, and a custom serve script is no longer
necessary (just `python3 -m http.server` in out/domain
does the job).