Replace clj-markdown with nextjournal/markdown
The latter has a very nice property of compiling the markdown to
an AST first, and then converting that either to Hiccup or to
plain text. This enables us to plug in server-side syntax
highlighting easily, and avoids having to parse the HTML
generated by clj-markdown altogether.

The Gemini support will require a rewrite, which will come in
due course. Commented out for the time being.