Remove nohighlight
The server-side syntax highlighting implementation has a nice benefit
of throwing an exception when highlight.js doesn't understand a
language. "nohighlight" isn't one (it was used to prevent
highlighting) and we don't need it anymore, so replace it by an
ordinary Markdown code block start (without declaring a language)
across all blog posts.