Use Markdown syntax for figures in crucial places
Turns out markdown-it, and by extension nextjournal.markdown, has a
syntax for providing _both_ a figcaption and an alt text for an image,
and these may differ. Eventually I want to use it everywhere, but for
now put it in some crucial places only – unlike clj-markdown,
nextjournal.markdown doesn't process Markdown embedded within HTML.

This fixes image captions with links in two posts on plblog.